Created folders css, images, fonts, index.html, style.CSS. Added text and image format locally
In the created index.html file, a semantically correct page was created for the beauty zone of the site.
navigation, headings, paragraphs with a description were created, a footer was created, and in the style.css file, the color and text format that is needed were added to the texts.